+++ ICD.Connect.Protocol.Mock/Properties/AssemblyInfo.cs
@@ -4,4 +4,4 @@ using System.Reflection;
 [assembly: AssemblyCompany("ICD Systems")]
 [assembly: AssemblyProduct("ICD.Connect.Protocol.Mock")]
 [assembly: AssemblyCopyright("Copyright © ICD Systems 2020")]
-[assembly: AssemblyVersion("12.0.0.0")]
+[assembly: AssemblyVersion("12.0.1.0")]