+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+ - Crestron TCPServer now spawns a new thread to handle SocketStatusChanged event - mitigation for TCPServer bugs
 
 ## [12.0.0] - 2020-03-20
 ### Added