+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
  - Crestron TCPServer now spawns a new thread to handle SocketStatusChanged event - mitigation for TCPServer bugs
+ - Crestron TCPServer - Fixed issue where connected but not listening would be interpreted as listening
+ - Crestron TCPServer - Better handling and recovery when hitting max number of connections - will now allow additional connections when no longer at max
 
 ## [12.0.0] - 2020-03-20
 ### Added